% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getseeds.R
\name{getseeds}
\alias{getseeds}
\title{Set seeds for multiple runs}
\usage{
getseeds(N, seed)
}
\arguments{
\item{N}{No. runs}

\item{seed}{Starting seed}
}
\value{
List of N sets of 7 seeds
}
\description{
Set seeds for multiple runs
}
